+++ lib/logreader.c
@@ -25,7 +25,7 @@
 #include "logreader.h"
 #include "mainloop-io-worker.h"
 #include "mainloop-call.h"
-#include "ack_tracker.h"
+#include "ack-tracker/ack_tracker.h"
 #include "scratch-buffers.h"
 
 #include <iv_event.h>
