+++ modules/systemd-journal/journal-reader.c
@@ -28,7 +28,7 @@
 #include "mainloop-io-worker.h"
 #include "persist-state.h"
 #include "persistable-state-header.h"
-#include "ack_tracker.h"
+#include "ack-tracker/ack_tracker.h"
 #include "parse-number.h"
 #include "journal-reader.h"
 #include "timeutils/misc.h"
